[#179193861] Roughly setup components to render different UI based on mobile/desktop
