+++ src/components/basins-tab/filters/mobile/index.js
@@ -1,0 +1,3 @@
+import MobileFilters from './component';
+
+export default MobileFilters;
